4/12/24 2:02 PM
Changed evaluation values to use i16 instead of i32

Added zobrist make_move and undo_move

Implemented uci commands: uci, setoption Hash, isready, ucinewgame, and quit

Added go "fulltime" and "depth"

Finished implementing the transposition table

Implemented iterative deepening

Updated TODO list

+++ src/move_list.rs
@@ -1,7 +1,7 @@
 use crate::move_data::MoveData;
 
 pub struct MoveList {
-	pub moves: Vec<(MoveData, i32)>,
+	pub moves: Vec<(MoveData, i16)>,
 }
 
 impl MoveList {
@@ -19,7 +19,7 @@ impl MoveList {
 		self.moves.push((value, 0));
 	}
 
-	pub fn pop(&mut self) -> Option<(MoveData, i32)> {
+	pub fn pop(&mut self) -> Option<(MoveData, i16)> {
 		self.moves.pop()
 	}
